requirements.txt
# COALESCE: Cost-Optimized Agent Labor Exchange via Skill-based Competence Estimation
# Requirements for running the simulation

# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Configuration
PyYAML>=6.0

# Data processing
dataclasses>=0.6

# Optional dependencies for enhanced functionality
# jupyter>=1.0.0  # For notebook analysis
# plotly>=5.0.0   # For interactive charts